requirements.txt
# AI/LLM
anthropic>=0.18.0
openai>=1.0.0
langgraph>=0.2.0
langchain>=0.2.0

# API clients
requests>=2.31.0
httpx>=0.25.0
tavily-python>=0.3.0

# Environment management
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0
PyYAML>=6.0.0

# Text matching
rapidfuzz>=3.0.0

# Google Sheets API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
